+++ lib/src/ast/selector/id.dart
@@ -5,6 +5,7 @@
 import 'dart:math' as math;
 
 import 'package:meta/meta.dart';
+import 'package:source_span/source_span.dart';
 
 import '../../visitor/interface/selector.dart';
 import '../selector.dart';
@@ -21,13 +22,13 @@ class IDSelector extends SimpleSelector {
 
   int get specificity => math.pow(super.specificity, 2) as int;
 
-  IDSelector(this.name);
+  IDSelector(this.name, FileSpan span) : super(span);
 
   T accept<T>(SelectorVisitor<T> visitor) => visitor.visitIDSelector(this);
 
   /// @nodoc
   @internal
-  IDSelector addSuffix(String suffix) => IDSelector(name + suffix);
+  IDSelector addSuffix(String suffix) => IDSelector(name + suffix, span);
 
   /// @nodoc
   @internal