+++ lib/src/ast/selector/parent.dart
@@ -3,6 +3,7 @@
 // https://opensource.org/licenses/MIT.
 
 import 'package:meta/meta.dart';
+import 'package:source_span/source_span.dart';
 
 import '../../visitor/interface/selector.dart';
 import '../selector.dart';
@@ -22,7 +23,7 @@ class ParentSelector extends SimpleSelector {
   /// indicating that the parent selector will not be modified.
   final String? suffix;
 
-  ParentSelector({this.suffix});
+  ParentSelector(FileSpan span, {this.suffix}) : super(span);
 
   T accept<T>(SelectorVisitor<T> visitor) => visitor.visitParentSelector(this);