+++ lib/src/utils.dart
@@ -464,6 +464,16 @@ extension MapExtension<K, V> on Map<K, V> {
 }
 
 extension IterableExtension<E> on Iterable<E> {
+  /// Returns the first `T` returned by [callback] for an element of [iterable],
+  /// or `null` if it returns `null` for every element.
+  T? search<T>(T? Function(E element) callback) {
+    for (var element in this) {
+      var value = callback(element);
+      if (value != null) return value;
+    }
+    return null;
+  }
+
   /// Returns a view of this list that covers all elements except the last.
   ///
   /// Note this is only efficient for an iterable with a known length.