+++ lib/src/visitor/async_evaluate.dart
@@ -31,6 +31,7 @@ import '../functions.dart';
 import '../functions/meta.dart' as meta;
 import '../importer.dart';
 import '../importer/legacy_node.dart';
+import '../interpolation_map.dart';
 import '../io.dart';
 import '../logger.dart';
 import '../module.dart';
@@ -40,6 +41,7 @@ import '../syntax.dart';
 import '../utils.dart';
 import '../util/multi_span.dart';
 import '../util/nullable.dart';
+import '../util/span.dart';
 import '../value.dart';
 import 'expression_to_calc.dart';
 import 'interface/css.dart';
@@ -525,7 +527,7 @@ class _EvaluateVisitor
         if (!(_asNodeSass && url.toString() == 'stdin')) _loadedUrls.add(url);
       }
 
-      var module = await _execute(importer, node);
+      var module = await _addExceptionTrace(() => _execute(importer, node));
 
       return EvaluateResult(_combineCss(module), _loadedUrls);
     });
@@ -534,14 +536,14 @@ class _EvaluateVisitor
   Future<Value> runExpression(AsyncImporter? importer, Expression expression) =>
       withEvaluationContext(
           _EvaluationContext(this, expression),
-          () => _withFakeStylesheet(
-              importer, expression, () => expression.accept(this)));
+          () => _withFakeStylesheet(importer, expression,
+              () => _addExceptionTrace(() => expression.accept(this))));
 
   Future<void> runStatement(AsyncImporter? importer, Statement statement) =>
       withEvaluationContext(
           _EvaluationContext(this, statement),
-          () => _withFakeStylesheet(
-              importer, statement, () => statement.accept(this)));
+          () => _withFakeStylesheet(importer, statement,
+              () => _addExceptionTrace(() => statement.accept(this))));
 
   /// Asserts that [value] is not `null` and returns it.
   ///
@@ -638,29 +640,8 @@ class _EvaluateVisitor
         _inDependency = oldInDependency;
       }
 
-      try {
-        await callback(module);
-      } on SassRuntimeException {
-        rethrow;
-      } on MultiSpanSassException catch (error, stackTrace) {
-        throwWithTrace(
-            MultiSpanSassRuntimeException(
-                error.message,
-                error.span,
-                error.primaryLabel,
-                error.secondarySpans,
-                _stackTrace(error.span)),
-            stackTrace);
-      } on SassException catch (error, stackTrace) {
-        throwWithTrace(_exception(error.message, error.span), stackTrace);
-      } on MultiSpanSassScriptException catch (error, stackTrace) {
-        throwWithTrace(
-            _multiSpanException(
-                error.message, error.primaryLabel, error.secondarySpans),
-            stackTrace);
-      } on SassScriptException catch (error, stackTrace) {
-        throwWithTrace(_exception(error.message), stackTrace);
-      }
+      await _addExceptionSpanAsync(nodeWithSpan, () => callback(module),
+          addStackFrame: false);
     });
   }
 
@@ -940,10 +921,12 @@ class _EvaluateVisitor
     var query = AtRootQuery.defaultQuery;
     var unparsedQuery = node.query;
     if (unparsedQuery != null) {
-      var resolved =
-          await _performInterpolation(unparsedQuery, warnForColor: true);
-      query = _adjustParseError(
-          unparsedQuery, () => AtRootQuery.parse(resolved, logger: _logger));
+      var tuple =
+          await _performInterpolationWithMap(unparsedQuery, warnForColor: true);
+      var resolved = tuple.item1;
+      var map = tuple.item2;
+      query =
+          AtRootQuery.parse(resolved, interpolationMap: map, logger: _logger);
     }
 
     var parent = _parent;
@@ -1220,20 +1203,18 @@ class _EvaluateVisitor
                   'This will be an error in Dart Sass 2.0.0.\n'
                   '\n'
                   'More info: https://sass-lang.com/d/bogus-combinators',
-          MultiSpan(styleRule.selector.span, 'invalid selector',
+          MultiSpan(complex.span.trimRight(), 'invalid selector',
               {node.span: '@extend rule'}),
           deprecation: true);
     }
 
-    var targetText =
-        await _interpolationToValue(node.selector, warnForColor: true);
+    var tuple =
+        await _performInterpolationWithMap(node.selector, warnForColor: true);
+    var targetText = tuple.item1;
+    var targetMap = tuple.item2;
 
-    var list = _adjustParseError(
-        targetText,
-        () => SelectorList.parse(
-            trimAscii(targetText.value, excludeEscape: true),
-            logger: _logger,
-            allowParent: false));
+    var list = SelectorList.parse(trimAscii(targetText, excludeEscape: true),
+        interpolationMap: targetMap, logger: _logger, allowParent: false);
 
     for (var complex in list.components) {
       var compound = complex.singleCompound;
@@ -1241,7 +1222,7 @@ class _EvaluateVisitor
         // If the selector was a compound selector but not a simple
         // selector, emit a more explicit error.
         throw SassFormatException(
-            "complex selectors may not be extended.", targetText.span);
+            "complex selectors may not be extended.", complex.span);
       }
 
       var simple = compound.singleSimple;
@@ -1250,7 +1231,7 @@ class _EvaluateVisitor
             "compound selectors may no longer be extended.\n"
             "Consider `@extend ${compound.components.join(', ')}` instead.\n"
             "See https://sass-lang.com/d/extend-compound for details.\n",
-            targetText.span);
+            compound.span);
       }
 
       _extensionStore.addExtension(
@@ -1651,8 +1632,8 @@ class _EvaluateVisitor
       } else {
         throw "Can't find stylesheet to import.";
       }
-    } on SassException catch (error, stackTrace) {
-      throwWithTrace(_exception(error.message, error.span), stackTrace);
+    } on SassException {
+      rethrow;
     } on ArgumentError catch (error, stackTrace) {
       throwWithTrace(_exception(error.toString()), stackTrace);
     } catch (error, stackTrace) {
@@ -1837,12 +1818,12 @@ class _EvaluateVisitor
   /// queries.
   Future<List<CssMediaQuery>> _visitMediaQueries(
       Interpolation interpolation) async {
-    var resolved =
-        await _performInterpolation(interpolation, warnForColor: true);
-
-    // TODO(nweiz): Remove this type argument when sdk#31398 is fixed.
-    return _adjustParseError<List<CssMediaQuery>>(interpolation,
-        () => CssMediaQuery.parseList(resolved, logger: _logger));
+    var tuple =
+        await _performInterpolationWithMap(interpolation, warnForColor: true);
+    var resolved = tuple.item1;
+    var map = tuple.item2;
+    return CssMediaQuery.parseList(resolved,
+        logger: _logger, interpolationMap: map);
   }
 
   /// Returns a list of queries that selects for contexts that match both
@@ -1879,16 +1860,18 @@ class _EvaluateVisitor
           "Style rules may not be used within nested declarations.", node.span);
     }
 
-    var selectorText = await _interpolationToValue(node.selector,
-        trim: true, warnForColor: true);
+    var tuple =
+        await _performInterpolationWithMap(node.selector, warnForColor: true);
+    var selectorText = tuple.item1;
+    var selectorMap = tuple.item2;
+
     if (_inKeyframes) {
       // NOTE: this logic is largely duplicated in [visitCssKeyframeBlock]. Most
       // changes here should be mirrored there.
 
-      var parsedSelector = _adjustParseError(
-          node.selector,
-          () => KeyframeSelectorParser(selectorText.value, logger: _logger)
-              .parse());
+      var parsedSelector = KeyframeSelectorParser(selectorText,
+              logger: _logger, interpolationMap: selectorMap)
+          .parse();
       var rule = ModifiableCssKeyframeBlock(
           CssValue(List.unmodifiable(parsedSelector), node.selector.span),
           node.span);
@@ -1902,20 +1885,15 @@ class _EvaluateVisitor
       return null;
     }
 
-    var parsedSelector = _adjustParseError(
-        node.selector,
-        () => SelectorList.parse(selectorText.value,
+    var parsedSelector = SelectorList.parse(selectorText,
+            interpolationMap: selectorMap,
             allowParent: !_stylesheet.plainCss,
             allowPlaceholder: !_stylesheet.plainCss,
-            logger: _logger));
-    parsedSelector = _addExceptionSpan(
-        node.selector,
-        () => parsedSelector.resolveParentSelectors(
-            _styleRuleIgnoringAtRoot?.originalSelector,
-            implicitParent: !_atRootExcludingStyleRule));
+            logger: _logger)
+        .resolveParentSelectors(_styleRuleIgnoringAtRoot?.originalSelector,
+            implicitParent: !_atRootExcludingStyleRule);
 
-    var selector = _extensionStore.addSelector(
-        parsedSelector, node.selector.span, _mediaQueries);
+    var selector = _extensionStore.addSelector(parsedSelector, _mediaQueries);
     var rule = ModifiableCssStyleRule(selector, node.span,
         originalSelector: parsedSelector);
     var oldAtRootExcludingStyleRule = _atRootExcludingStyleRule;
@@ -1942,7 +1920,7 @@ class _EvaluateVisitor
               'This will be an error in Dart Sass 2.0.0.\n'
               '\n'
               'More info: https://sass-lang.com/d/bogus-combinators',
-              node.selector.span,
+              complex.span.trimRight(),
               deprecation: true);
         } else if (complex.leadingCombinators.isNotEmpty) {
           _warn(
@@ -1950,7 +1928,7 @@ class _EvaluateVisitor
               'This will be an error in Dart Sass 2.0.0.\n'
               '\n'
               'More info: https://sass-lang.com/d/bogus-combinators',
-              node.selector.span,
+              complex.span.trimRight(),
               deprecation: true);
         } else {
           _warn(
@@ -1964,7 +1942,7 @@ class _EvaluateVisitor
                       'This will be an error in Dart Sass 2.0.0.\n'
                       '\n'
                       'More info: https://sass-lang.com/d/bogus-combinators',
-              MultiSpan(node.selector.span, 'invalid selector', {
+              MultiSpan(complex.span.trimRight(), 'invalid selector', {
                 rule.children.first.span: "this is not a style rule" +
                     (rule.children.every((child) => child is CssComment)
                         ? '\n(try converting to a //-style comment)'
@@ -2703,27 +2681,10 @@ class _EvaluateVisitor
 
     Value result;
     try {
-      result = await callback(evaluated.positional);
-    } on SassRuntimeException {
+      result = await _addExceptionSpanAsync(
+          nodeWithSpan, () => callback(evaluated.positional));
+    } on SassException {
       rethrow;
-    } on MultiSpanSassScriptException catch (error, stackTrace) {
-      throwWithTrace(
-          MultiSpanSassRuntimeException(
-              error.message,
-              nodeWithSpan.span,
-              error.primaryLabel,
-              error.secondarySpans,
-              _stackTrace(nodeWithSpan.span)),
-          stackTrace);
-    } on MultiSpanSassException catch (error, stackTrace) {
-      throwWithTrace(
-          MultiSpanSassRuntimeException(
-              error.message,
-              error.span,
-              error.primaryLabel,
-              error.secondarySpans,
-              _stackTrace(error.span)),
-          stackTrace);
     } catch (error, stackTrace) {
       String? message;
       try {
@@ -3100,11 +3061,10 @@ class _EvaluateVisitor
     }
 
     var styleRule = _styleRule;
-    var originalSelector = node.selector.value.resolveParentSelectors(
+    var originalSelector = node.selector.resolveParentSelectors(
         styleRule?.originalSelector,
         implicitParent: !_atRootExcludingStyleRule);
-    var selector = _extensionStore.addSelector(
-        originalSelector, node.selector.span, _mediaQueries);
+    var selector = _extensionStore.addSelector(originalSelector, _mediaQueries);
     var rule = ModifiableCssStyleRule(selector, node.span,
         originalSelector: originalSelector);
     var oldAtRootExcludingStyleRule = _atRootExcludingStyleRule;
@@ -3206,10 +3166,42 @@ class _EvaluateVisitor
   /// values passed into the interpolation.
   Future<String> _performInterpolation(Interpolation interpolation,
       {bool warnForColor = false}) async {
+    var tuple = await _performInterpolationHelper(interpolation,
+        sourceMap: true, warnForColor: warnForColor);
+    return tuple.item1;
+  }
+
+  /// Like [_performInterpolation], but also returns a [InterpolationMap] that
+  /// can map spans from the resulting string back to the original
+  /// [interpolation].
+  Future<Tuple2<String, InterpolationMap>> _performInterpolationWithMap(
+      Interpolation interpolation,
+      {bool warnForColor = false}) async {
+    var tuple = await _performInterpolationHelper(interpolation,
+        sourceMap: true, warnForColor: warnForColor);
+    return Tuple2(tuple.item1, tuple.item2!);
+  }
+
+  /// A helper that implements the core logic of both [_performInterpolation]
+  /// and [_performInterpolationWithMap].
+  Future<Tuple2<String, InterpolationMap?>> _performInterpolationHelper(
+      Interpolation interpolation,
+      {required bool sourceMap,
+      bool warnForColor = false}) async {
+    var targetLocations = sourceMap ? <SourceLocation>[] : null;
     var oldInSupportsDeclaration = _inSupportsDeclaration;
     _inSupportsDeclaration = false;
-    var result = (await mapAsync(interpolation.contents, (value) async {
-      if (value is String) return value;
+    var buffer = StringBuffer();
+    var first = true;
+    for (var value in interpolation.contents) {
+      if (!first) targetLocations?.add(SourceLocation(buffer.length));
+      first = false;
+
+      if (value is String) {
+        buffer.write(value);
+        continue;
+      }
+
       var expression = value as Expression;
       var result = await expression.accept(this);
 
@@ -3232,11 +3224,15 @@ class _EvaluateVisitor
             expression.span);
       }
 
-      return _serialize(result, expression, quote: false);
-    }))
-        .join();
+      buffer.write(_serialize(result, expression, quote: false));
+    }
     _inSupportsDeclaration = oldInSupportsDeclaration;
-    return result;
+
+    return Tuple2(
+        buffer.toString(),
+        targetLocations == null
+            ? null
+            : InterpolationMap(interpolation, targetLocations));
   }
 
   /// Evaluates [expression] and calls `toCssString()` and wraps a
@@ -3452,73 +3448,53 @@ class _EvaluateVisitor
       MultiSpanSassRuntimeException(message, _stack.last.item2.span,
           primaryLabel, secondaryLabels, _stackTrace());
 
-  /// Runs [callback], and adjusts any [SassFormatException] to be within
-  /// [nodeWithSpan]'s source span.
-  ///
-  /// Specifically, this adjusts format exceptions so that the errors are
-  /// reported as though the text being parsed were exactly in [span]. This may
-  /// not be quite accurate if the source text contained interpolation, but
-  /// it'll still produce a useful error.
-  ///
-  /// This takes an [AstNode] rather than a [FileSpan] so it can avoid calling
-  /// [AstNode.span] if the span isn't required, since some nodes need to do
-  /// real work to manufacture a source span.
-  T _adjustParseError<T>(AstNode nodeWithSpan, T callback()) {
-    try {
-      return callback();
-    } on SassFormatException catch (error, stackTrace) {
-      var errorText = error.span.file.getText(0);
-      var span = nodeWithSpan.span;
-      var syntheticFile = span.file
-          .getText(0)
-          .replaceRange(span.start.offset, span.end.offset, errorText);
-      var syntheticSpan =
-          SourceFile.fromString(syntheticFile, url: span.file.url).span(
-              span.start.offset + error.span.start.offset,
-              span.start.offset + error.span.end.offset);
-      throwWithTrace(_exception(error.message, syntheticSpan), stackTrace);
-    }
-  }
-
   /// Runs [callback], and converts any [SassScriptException]s it throws to
   /// [SassRuntimeException]s with [nodeWithSpan]'s source span.
   ///
   /// This takes an [AstNode] rather than a [FileSpan] so it can avoid calling
   /// [AstNode.span] if the span isn't required, since some nodes need to do
   /// real work to manufacture a source span.
-  T _addExceptionSpan<T>(AstNode nodeWithSpan, T callback()) {
+  ///
+  /// If [addStackFrame] is true (the default), this will add an innermost stack
+  /// frame for [nodeWithSpan]. Otherwise, it will use the existing stack as-is.
+  T _addExceptionSpan<T>(AstNode nodeWithSpan, T callback(),
+      {bool addStackFrame = true}) {
     try {
       return callback();
-    } on MultiSpanSassScriptException catch (error, stackTrace) {
-      throwWithTrace(
-          MultiSpanSassRuntimeException(
-              error.message,
-              nodeWithSpan.span,
-              error.primaryLabel,
-              error.secondarySpans,
-              _stackTrace(nodeWithSpan.span)),
-          stackTrace);
     } on SassScriptException catch (error, stackTrace) {
-      throwWithTrace(_exception(error.message, nodeWithSpan.span), stackTrace);
+      throwWithTrace(
+          error
+              .withSpan(nodeWithSpan.span)
+              .withTrace(_stackTrace(addStackFrame ? nodeWithSpan.span : null)),
+          stackTrace);
     }
   }
 
   /// Like [_addExceptionSpan], but for an asynchronous [callback].
   Future<T> _addExceptionSpanAsync<T>(
-      AstNode nodeWithSpan, FutureOr<T> callback()) async {
+      AstNode nodeWithSpan, FutureOr<T> callback(),
+      {bool addStackFrame = true}) async {
     try {
       return await callback();
-    } on MultiSpanSassScriptException catch (error, stackTrace) {
-      throwWithTrace(
-          MultiSpanSassRuntimeException(
-              error.message,
-              nodeWithSpan.span,
-              error.primaryLabel,
-              error.secondarySpans,
-              _stackTrace(nodeWithSpan.span)),
-          stackTrace);
     } on SassScriptException catch (error, stackTrace) {
-      throwWithTrace(_exception(error.message, nodeWithSpan.span), stackTrace);
+      throwWithTrace(
+          error
+              .withSpan(nodeWithSpan.span)
+              .withTrace(_stackTrace(addStackFrame ? nodeWithSpan.span : null)),
+          stackTrace);
+    }
+  }
+
+  /// Runs [callback], and converts any [SassException]s that aren't already
+  /// [SassRuntimeException]s to [SassRuntimeException]s with the current stack
+  /// trace.
+  Future<T> _addExceptionTrace<T>(FutureOr<T> callback()) async {
+    try {
+      return await callback();
+    } on SassRuntimeException {
+      rethrow;
+    } on SassException catch (error, stackTrace) {
+      throwWithTrace(error.withTrace(_stackTrace(error.span)), stackTrace);
     }
   }