+++ pkg/sass_api/lib/sass_api.dart
@@ -26,6 +26,7 @@ export 'package:sass/src/visitor/recursive_ast.dart';
 export 'package:sass/src/visitor/recursive_selector.dart';
 export 'package:sass/src/visitor/recursive_statement.dart';
 export 'package:sass/src/visitor/replace_expression.dart';
+export 'package:sass/src/visitor/selector_search.dart';
 export 'package:sass/src/visitor/statement_search.dart';
 
 /// Parses [text] as a CSS identifier and returns the result.